+++ src/main/java/kg/peaksoft/peaksoftlmsbb4/api/AdminApi.java
@@ -30,16 +30,5 @@ public class AdminApi {
     private AdminResponse saveAdmin(@RequestBody AdminRequest adminRequest){
         return adminService.saveAdmin(adminRequest);
     }
-    @PostConstruct
-    private void init(){
-        Admin admin=new Admin();
-        admin.setFirstName("Rahim");
-        admin.setLastName("Kurbanov");
-        User user =new User();
-        user.setRole(Role.ADMIN);
-        user.setEmail("kurbanov@gmail.com");
-        user.setPassword(passwordEncoder.encode("123123"));
-        admin.setUser(user);
-        adminRepository.save(admin);
-    }
+
 }
