+++ src/main/java/kg/peaksoft/peaksoftlmsbb4/repository/AdminRepository.java
@@ -2,7 +2,6 @@ package kg.peaksoft.peaksoftlmsbb4.repository;
 
 import kg.peaksoft.peaksoftlmsbb4.model.Admin;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
@@ -10,6 +9,7 @@ import java.util.Optional;
 @Repository
 public interface AdminRepository extends JpaRepository<Admin, Long> {
     boolean existsByUser_Email(String email);
+
     Optional<Admin> findByUserEmail(String email);
 
 }
